+++ picwall/src/main/java/cn/rs/picwall/pic/dao/UserDao.java
@@ -2,12 +2,9 @@ package cn.rs.picwall.pic.dao;
 
 import cn.rs.picwall.pic.pojo.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.DeleteMapping;
 
-public interface UserDao extends JpaRepository<User, Integer> {
+public interface UserDao extends JpaRepository<User, Long> {
 
     User findByName(String name);
